Adding more documentation
